[SC-755] Change the 'nonce' type to Biginteger. Use scorex.util.StringTokenizer instead of java.util.StringTokenizer.

+++ java-score/src/main/java/com/iconloop/score/pds/DidMessage.java
@@ -1,12 +1,14 @@
 package com.iconloop.score.pds;
 
+import java.math.BigInteger;
+
 public class DidMessage {
     public final String did;
     public final String kid;
     public final String target;
-    public int nonce;
+    public BigInteger nonce;
 
-    public DidMessage(String did, String kid, String target, int nonce) {
+    public DidMessage(String did, String kid, String target, BigInteger nonce) {
         this.did = did;
         this.kid = kid;
         this.target = target;
